Properly make variables local
The function make-variable-buffer-local is used for declaring (at the
top level) that a certain variable should always be made buffer-local
if it is ever changed in any way. For explicitly making a variable
buffer-local, especially within a (set ...), one should use
make-local-variable, which makes a variable buffer-local for the
current buffer.

Emacs warns about this during byte-compilation.