Avoid indentation errors after 'end'.
This is, if not a fix, a workaround for half of #18. In my elixir files
this produces no indentation errors; maybe you can test this against
your setup and see if it doesn't raise any red flags for you.

I don't know enough smie (yet), to look after the statement indentation
error, that is the other error reported in #18.