Add a SMIE rule function for "def" tokens.
This function now explicitly sets the indentation level of lines
beginning with "def" to 2. Formerly, this token returned `nil` (the
default, because there was no matching pattern in `elixir-smie-rules`),
which caused SMIE to call `smie-indent-after-keyword` to calculate the
indentation. This function was returning the incorrect value which was
causing the inappropriate alignment to the `.`.