Remove grammar clause for "fn" terminal. Fixes #7
After stepping through SMIE functions with EDebug, I realized the only
actual problem in indenting blocks like these is that for the sake of
the grammar, `fn` is considered an opener. I tried many ways to get this
particular issue sorted out, but the simplest was just to remove that
clause from the grammar.