Fixed bug in test that was masking a problem being solved. Fixes #27
The indentation in the expected output was wrong (my fault) in a couple of ways. Who knows how long
this test should have been passing but wasn't.

Who cares though, this is the last known-fail test! This means this project is somewhat caught up now
on its backlog of issues.