Add fn for comparing line numbers.
Elixir is quite flexible wrt syntax around the `->` token. It can follow
`do` or `fn`, either on the same line or separately. Because of this, a
function that can detect whether it's on the same line as its parent was
needed.

After adding this function, all tests passed. I'm skeptical about the
robustness of this function but it will have to be battle hardened by
users reporting bugs.