Improve speed of smie by disabling features
This PR disables the worst offenders of lookbacks that make opening
large files very slow.

More testing is needed to determine if this is a worthwhile tradeoff.
The main reason it is palatable is that it fixes
https://github.com/elixir-editors/emacs-elixir/issues/463 and
indentation isn't as important as it once was because you can simply run
`elixir-format` (or `lsp-format-buffer`).